ran black, fixed formating, moved test to linearGaussian_snpe_test.
